+++ lib/src/widget/layout/column.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/widgets.dart';
+
+extension ColumnWidgetListExtension on List<Widget> {
+  Column column({
+    MainAxisAlignment mainAxisAlignment = MainAxisAlignment.start,
+    MainAxisSize mainAxisSize = MainAxisSize.max,
+    CrossAxisAlignment crossAxisAlignment = CrossAxisAlignment.center,
+    TextDirection? textDirection,
+    VerticalDirection verticalDirection = VerticalDirection.down,
+    double spacing = 0,
+    TextBaseline? textBaseline,
+  }) {
+    return Column(
+      mainAxisAlignment: mainAxisAlignment,
+      crossAxisAlignment: crossAxisAlignment,
+      mainAxisSize: mainAxisSize,
+      textDirection: textDirection,
+      verticalDirection: verticalDirection,
+      textBaseline: textBaseline,
+      spacing: spacing,
+      children: this,
+    );
+  }
+}
+
+extension ColumnWidgetExtension on Widget {
+  Column column({
+    MainAxisAlignment mainAxisAlignment = MainAxisAlignment.start,
+    MainAxisSize mainAxisSize = MainAxisSize.max,
+    CrossAxisAlignment crossAxisAlignment = CrossAxisAlignment.center,
+    TextDirection? textDirection,
+    VerticalDirection verticalDirection = VerticalDirection.down,
+    double spacing = 0,
+    TextBaseline? textBaseline,
+  }) {
+    return Column(
+      children: [this],
+    );
+  }
+}